backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Settings and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<5.0.0

# Storage
minio>=7.2.3

# AI
anthropic>=0.18.0
httpx>=0.26.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Utilities
python-dateutil>=2.8.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
ruff>=0.2.0
mypy>=1.8.0
